+++ skyperious/__main__.py
@@ -1,0 +1,5 @@
+"""Package entry point."""
+from skyperious import main
+
+if "__main__" == __name__:
+    main.run()